Improved type reconstruction

+++ tests/branching_kernels/if_2/if_2_dcmpl.cl
@@ -3,7 +3,7 @@ void if_2(int x, __global int *data, int y)
 {
     uint var0;
     var0 = get_global_id(0);
-    if (0 == (int)var0) {
+    if (0 == var0) {
         var0 = get_global_id(1);
     }
     data[var0] = x;
