+++ tests/branching_kernels/if_else_0_labels/if_else_0_labels_dcmpl-gfx1010.cl
@@ -12,7 +12,7 @@ void if_else_0_labels(int x, __global int *data, int y)
     else {
         var3 = get_global_id(1);
         var2 = 1;
-        var0 = (get_global_id(1) * x) - (uint)y;
+        var0 = (get_global_id(1) * x) - y;
     }
     data[var2] = var0;
     data[var3] = x;
