+++ tests/branching_kernels/if_else_and_if_else_gfx/if_else_and_if_else_gfx_dcmpl-gfx1030.cl
@@ -12,7 +12,7 @@ void if_else_and_if_else_1(int x, __global int *data, int y)
     else {
         var1 = get_global_id(1);
         var3 = 1;
-        var5 = (get_global_id(1) * x) - (uint)y;
+        var5 = (get_global_id(1) * x) - y;
     }
     data[var3] = var5;
     data[get_global_id(2)] = min(x, y);
