+++ tests/branching_kernels/if_else_in_if_of_if_else_gfx/if_else_in_if_of_if_else_gfx_dcmpl-gfx1030.cl
@@ -11,14 +11,14 @@ void if_else_in_if_of_if_else_1(int x, __global int *data, int y)
     }
     else {
         var4 = get_global_id(1);
-        data[get_global_id(0)] = (get_global_id(1) * x) - (uint)y;
+        data[get_global_id(0)] = (get_global_id(1) * x) - y;
         if (x >= y) {
-            var8 = (ulong)y + get_global_id(1);
+            var8 = y + get_global_id(1);
             var6 = get_global_id(2);
         }
         else {
             var6 = get_global_id(2);
-            var8 = (ulong)x + get_global_id(2);
+            var8 = x + get_global_id(2);
         }
     }
     data[var6] = var8;
@@ -36,12 +36,12 @@ void if_else_in_if_of_if_else_2(int x, __global int *data, int y)
     }
     else {
         var5 = get_global_id(1);
-        data[get_global_id(0)] = (get_global_id(1) * x) - (uint)y;
+        data[get_global_id(0)] = (get_global_id(1) * x) - y;
         if (var5 >= get_global_id(2)) {
-            var12 = (ulong)y + get_global_id(1);
+            var12 = y + get_global_id(1);
         }
         else {
-            var12 = (ulong)x + get_global_id(2);
+            var12 = x + get_global_id(2);
             var5 = get_global_id(2);
         }
     }
